+++ certbot-nginx/certbot_nginx/tests/configurator_test.py
@@ -972,11 +972,12 @@ class InstallSslOptionsConfTest(util.NginxTest):
         """
         from certbot_nginx.constants import ALL_SSL_OPTIONS_HASHES
         import pkg_resources
-        all_files = [pkg_resources.resource_filename("certbot_nginx", x) for x in (
-          "options-ssl-nginx.conf",
-          "options-ssl-nginx-old.conf",
-          "options-ssl-nginx-tls12-only.conf"
-          )]
+        all_files = [
+            pkg_resources.resource_filename("certbot_nginx", os.path.join("tls_configs", x))
+            for x in ("options-ssl-nginx.conf",
+                      "options-ssl-nginx-old.conf",
+                      "options-ssl-nginx-tls12-only.conf")
+        ]
         self.assertTrue(all_files)
         for one_file in all_files:
             file_hash = crypto_util.sha256sum(one_file)