+++ CHANGELOG.md
@@ -43,7 +43,7 @@ More details about these changes can be found on our GitHub repo.
 
 ### Fixed
 
-*
+* Fix certbot-auto failures on RHEL 8.
 
 More details about these changes can be found on our GitHub repo.